Move internal/pkg/test/exec to internal/pkg/test/tool/exec
