+++ e2e/help/help.go
@@ -15,7 +15,7 @@ import (
 	"testing"
 
 	"github.com/sylabs/singularity/e2e/internal/e2e"
-	"github.com/sylabs/singularity/internal/pkg/test/exec"
+	"github.com/sylabs/singularity/internal/pkg/test/tool/exec"
 	"gotest.tools/assert"
 	"gotest.tools/golden"
 )
